+++ src/front-chat/front-chat.module.ts
@@ -4,12 +4,14 @@ import { SlackMessageClient } from 'src/api/slack/slack-api';
 import { ZapierWebhookClient } from 'src/api/zapier/zapier-webhook-client';
 import { AuthModule } from 'src/auth/auth.module';
 import { ChatUserEntity } from 'src/entities/chat-user.entity';
+import { EventLogEntity } from 'src/entities/event-log.entity';
 import { PartnerAccessEntity } from 'src/entities/partner-access.entity';
 import { PartnerEntity } from 'src/entities/partner.entity';
 import { SubscriptionUserEntity } from 'src/entities/subscription-user.entity';
 import { SubscriptionEntity } from 'src/entities/subscription.entity';
 import { TherapySessionEntity } from 'src/entities/therapy-session.entity';
 import { UserEntity } from 'src/entities/user.entity';
+import { EventLoggerService } from 'src/event-logger/event-logger.service';
 import { FirebaseModule } from 'src/firebase/firebase.module';
 import { PartnerAccessService } from 'src/partner-access/partner-access.service';
 import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
@@ -21,12 +23,14 @@ import { FrontChatController } from './front-chat.controller';
 import { FrontChatGateway } from './front-chat.gateway';
 import { FrontChatScheduler } from './front-chat.scheduler';
 import { FrontChatService } from './front-chat.service';
+import { FrontChatWebhookService } from './front-chat-webhook.service';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([
       UserEntity,
       ChatUserEntity,
+      EventLogEntity,
       PartnerAccessEntity,
       PartnerEntity,
       SubscriptionEntity,
@@ -40,6 +44,8 @@ import { FrontChatService } from './front-chat.service';
     FrontChatService,
     FrontChatGateway,
     FrontChatScheduler,
+    FrontChatWebhookService,
+    EventLoggerService,
     PartnerAccessService,
     UserService,
     ServiceUserProfilesService,
@@ -50,6 +56,6 @@ import { FrontChatService } from './front-chat.service';
     SlackMessageClient,
   ],
   controllers: [FrontChatController],
-  exports: [FrontChatService, FrontChatGateway],
+  exports: [FrontChatService, FrontChatGateway, FrontChatWebhookService],
 })
 export class FrontChatModule {}
