+++ src/webhooks/webhooks.service.spec.ts
@@ -6,7 +6,6 @@ import { SlackMessageClient } from 'src/api/slack/slack-api';
 import { CoursePartnerService } from 'src/course-partner/course-partner.service';
 import { CoursePartnerEntity } from 'src/entities/course-partner.entity';
 import { CourseEntity } from 'src/entities/course.entity';
-import { EventLogEntity } from 'src/entities/event-log.entity';
 import { PartnerAccessEntity } from 'src/entities/partner-access.entity';
 import { PartnerAdminEntity } from 'src/entities/partner-admin.entity';
 import { PartnerEntity } from 'src/entities/partner.entity';
@@ -14,10 +13,6 @@ import { ResourceEntity } from 'src/entities/resource.entity';
 import { SessionEntity } from 'src/entities/session.entity';
 import { TherapySessionEntity } from 'src/entities/therapy-session.entity';
 import { UserEntity } from 'src/entities/user.entity';
-import { EVENT_NAME } from 'src/event-logger/event-logger.interface';
-import { EventLoggerService } from 'src/event-logger/event-logger.service';
-import { FrontChatGateway } from 'src/front-chat/front-chat.gateway';
-import { FrontChatService } from 'src/front-chat/front-chat.service';
 import { PartnerService } from 'src/partner/partner.service';
 import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import {
@@ -42,7 +37,6 @@ import {
   mockCoursePartnerRepositoryMethods,
   mockCoursePartnerServiceMethods,
   mockCourseRepositoryMethods,
-  mockEventLoggerRepositoryMethods,
   mockPartnerAccessRepositoryMethods,
   mockPartnerAdminRepositoryMethods,
   mockPartnerRepositoryMethods,
@@ -110,11 +104,6 @@ describe('WebhooksService', () => {
     mockPartnerAdminRepositoryMethods,
   );
   const mockedServiceUserProfilesService = createMock<ServiceUserProfilesService>();
-  const mockFrontChatService = createMock<FrontChatService>();
-  const mockEventLoggerService = createMock<EventLoggerService>();
-  const mockEventLogRepository = createMock<Repository<EventLogEntity>>(
-    mockEventLoggerRepositoryMethods,
-  );
 
   beforeEach(async () => {
     jest.clearAllMocks();
@@ -173,16 +162,6 @@ describe('WebhooksService', () => {
           useValue: mockedSlackMessageClient,
         },
         PartnerService,
-        {
-          provide: getRepositoryToken(EventLogEntity),
-          useValue: mockEventLogRepository,
-        },
-        { provide: FrontChatService, useValue: mockFrontChatService },
-        { provide: EventLoggerService, useValue: mockEventLoggerService },
-        {
-          provide: FrontChatGateway,
-          useValue: { emitAgentReply: jest.fn() },
-        },
       ],
     }).compile();
 
@@ -849,85 +828,4 @@ describe('WebhooksService', () => {
     });
   });
 
-  describe('handleFrontChatWebhook', () => {
-    const inboundPayload = {
-      id: 'evt_123',
-      type: 'inbound',
-      emitted_at: 1700000000,
-      conversation: {
-        id: 'cnv_abc',
-        recipient: { handle: 'user@example.com', role: 'to' },
-      },
-    };
-
-    beforeEach(() => {
-      jest.spyOn(mockFrontChatService, 'updateChatUserByEmail').mockResolvedValue(null);
-    });
-
-    it('should log CHAT_MESSAGE_SENT for inbound events', async () => {
-      await service.handleFrontChatWebhook(inboundPayload);
-
-      expect(mockEventLoggerService.createEventLog).toHaveBeenCalledWith(
-        { event: EVENT_NAME.CHAT_MESSAGE_SENT, date: new Date(1700000000 * 1000) },
-        'user@example.com',
-      );
-    });
-
-    it('should log CHAT_MESSAGE_RECEIVED for outbound events', async () => {
-      await service.handleFrontChatWebhook({ ...inboundPayload, type: 'outbound' });
-
-      expect(mockEventLoggerService.createEventLog).toHaveBeenCalledWith(
-        { event: EVENT_NAME.CHAT_MESSAGE_RECEIVED, date: new Date(1700000000 * 1000) },
-        'user@example.com',
-      );
-    });
-
-    it('should log CHAT_MESSAGE_RECEIVED for out_reply events', async () => {
-      await service.handleFrontChatWebhook({ ...inboundPayload, type: 'out_reply' });
-
-      expect(mockEventLoggerService.createEventLog).toHaveBeenCalledWith(
-        { event: EVENT_NAME.CHAT_MESSAGE_RECEIVED, date: expect.any(Date) },
-        'user@example.com',
-      );
-    });
-
-    it('should skip untracked event types', async () => {
-      await service.handleFrontChatWebhook({ ...inboundPayload, type: 'tag' });
-      expect(mockEventLoggerService.createEventLog).not.toHaveBeenCalled();
-    });
-
-    it('should skip events with no recipient email', async () => {
-      await service.handleFrontChatWebhook({
-        ...inboundPayload,
-        conversation: { id: 'cnv_abc' },
-      });
-      expect(mockEventLoggerService.createEventLog).not.toHaveBeenCalled();
-    });
-
-    it('should not throw if event logging fails', async () => {
-      mockEventLoggerService.createEventLog.mockRejectedValueOnce(new Error('User not found'));
-
-      await expect(service.handleFrontChatWebhook(inboundPayload)).resolves.not.toThrow();
-    });
-
-    it('saves conversation ID from inbound events for history loading on reconnect', async () => {
-      await service.handleFrontChatWebhook(inboundPayload);
-      await new Promise((r) => setImmediate(r));
-
-      expect(mockFrontChatService.updateChatUserByEmail).toHaveBeenCalledWith(
-        'user@example.com',
-        { frontConversationId: 'cnv_abc' },
-      );
-    });
-
-    it('does not try to save conversation ID from inbound events when conversation is absent', async () => {
-      await service.handleFrontChatWebhook({ ...inboundPayload, conversation: undefined });
-      await new Promise((r) => setImmediate(r));
-
-      expect(mockFrontChatService.updateChatUserByEmail).not.toHaveBeenCalledWith(
-        expect.anything(),
-        expect.objectContaining({ frontConversationId: expect.anything() }),
-      );
-    });
-  });
 });
